adjust namespaces
