+++ AlastairLundy.Extensions.Settings/AlastairLundy.Extensions.Settings/StoreProviders/CachedTextFileStoreProvider.cs
@@ -9,14 +9,16 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
-using System.Text;
 using System.Threading.Tasks;
+using AlastairLundy.Extensions.Settings.Internal;
+using AlastairLundy.Extensions.Settings.StoreProviders.Abstractions;
 
-using AlastairLundy.Extensions.Settings.Stores.Abstractions;
 // ReSharper disable RedundantExtendsListEntry
+// ReSharper disable NullableWarningSuppressionIsUsed
 
-namespace AlastairLundy.Extensions.Settings.Stores;
+namespace AlastairLundy.Extensions.Settings.StoreProviders;
 
 /// <summary>
 /// A text file based settings store with caching.
@@ -34,15 +36,12 @@ public class CachedTextFileStoreProvider<TValue> : TextFileStoreProvider<TValue>
     /// 
     /// </summary>
     /// <param name="fileConfiguration"></param>
-    /// <param name="toTValueFunc"></param>
-    /// <param name="toStringFunc"></param>
+    /// <param name="typeConverter"></param>
     /// <param name="keyValueSeparator"></param>
-    public CachedTextFileSettingsStore(FileStoreConfiguration fileConfiguration,
-        Func<string, TValue> toTValueFunc,
-        Func<TValue, string> toStringFunc,
+    public CachedTextFileStoreProvider(FileStoreConfiguration fileConfiguration,
+        TypeConverter typeConverter,
         char keyValueSeparator = '=') : base(fileConfiguration,
-        toTValueFunc,
-        toStringFunc,
+        typeConverter,
         keyValueSeparator)
     {
         Cache = new Dictionary<string, TValue>();
@@ -76,10 +75,18 @@ public class CachedTextFileStoreProvider<TValue> : TextFileStoreProvider<TValue>
         foreach (string line in lines)
         {
             string[] parts = line.Split(_keyValueSeparator);
+
+            if (Converter.CanConvertTo(typeof(TValue)) && Converter.CanConvertFrom(typeof(string)))
+            {
+                TValue value = (TValue)Converter.ConvertFromString(parts[1])!;
                 
-            TValue value = ToTValueConverter(parts[1]);
-                
-            Cache[key: parts[0]] = value;
+                Cache[key: parts[0]] = value;
+            }
+            else
+            {
+                throw new ArgumentException(
+                    Resources.Exceptions_Conversions_CannotConvertFromString.Replace("{x}", nameof(TValue)));
+            }
         }
         
         CacheLifetime = expiry;
