+++ AlastairLundy.Extensions.Settings/AlastairLundy.Extensions.Settings/StoreProviders/TextFileStoreProvider.cs
@@ -9,6 +9,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
 using System.Linq;
 using System.Text;
